requirement.txt
Django==1.11
Pillow==2.9.0
argparse==1.2.1
djangorestframework==3.6.2
pytz==2017.2
uuid==1.30
wsgiref==0.1.2
